# lib/api-zod/src/generated/types/batchDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Al-Saawi Pharmacy Drug Batch Management System
 * OpenAPI spec version: 0.1.0
 */

export interface BatchDetail {
  batch_id: number;
  batch_name: string;
  is_complete: boolean;
  persons_count: number;
  complete_count: number;
  total_cost_before: number;
  total_cost_after: number;
}