# lib/api-zod/src/generated/types/importRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Al-Saawi Pharmacy Drug Batch Management System
 * OpenAPI spec version: 0.1.0
 */
import type { ImportRequestDataItem } from "./importRequestDataItem";

export interface ImportRequest {
  data: ImportRequestDataItem[];
}