# lib/api-zod/src/generated/types/person.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Al-Saawi Pharmacy Drug Batch Management System
 * OpenAPI spec version: 0.1.0
 */

export interface Person {
  person_id: number;
  full_name: string;
  ph_number: string;
  location: string;
  batch_id?: number;
  is_complete: boolean;
  total_cost: number;
  inv_number: string;
  ordered_count: number;
  prepared_count: number;
  ready_count: number;
  total_drugs: number;
}