# lib/api-zod/src/generated/types/record.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Al-Saawi Pharmacy Drug Batch Management System
 * OpenAPI spec version: 0.1.0
 */

export interface Record {
  record_id: number;
  drug_id?: number;
  person_id?: number;
  batch_id?: number;
  final_price: number;
  ordered: boolean;
  prepared: boolean;
  ready: boolean;
  created_at: string;
}