# lib/api-zod/src/generated/types/updatePersonRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Al-Saawi Pharmacy Drug Batch Management System
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePersonRequest {
  full_name?: string;
  ph_number?: string;
  location?: string;
  inv_number?: string;
}